requirements.txt

telethon
python-decouple
aiohttp
